code refactor: 1.preparation for separation of normflowpy into a package

+++ analysis/analysis_helpers.py
@@ -11,12 +11,12 @@ def get_data_model(in_config):
     def generate_model_parameter_dict() -> dict:
         return {constants.DIM: in_config[constants.DIM],
                 constants.THETA_MIN: in_config[constants.THETA_MIN],
-                constants.THETA_DIM: in_config[constants.THETA_DIM],
+                constants.THETA_DIM: in_config.get(constants.THETA_DIM,1),
                 constants.SIGMA_N: in_config[constants.SIGMA_N],
                 constants.THETA_MAX: in_config[constants.THETA_MAX]}
 
     model_type = data_model.ModelType[in_config["model_type"].split(".")[-1]]
-    return data_model.get_model(model_type, generate_model_parameter_dict())
+    return data_model.get_model(model_type, generate_model_parameter_dict()), model_type
 
 
 def load_wandb_run(run_name):
@@ -30,17 +30,18 @@ def load_wandb_run(run_name):
             run.file("flow_best.pt").download()
 
             config = run.config
-            model_flow = generate_flow_model(config['dim'], config.get("theta_dim"), config['n_flow_blocks'],
+            model_flow = generate_flow_model(config['dim'], config.get("theta_dim",1), config['n_flow_blocks'],
                                              config["spline_flow"],
                                              n_layer_cond=config["n_layer_cond"],
                                              hidden_size_cond=config["hidden_size_cond"])
             model_flow.load_state_dict(torch.load(f"flow_best.pt", map_location=torch.device('cpu')))
             model_flow = model_flow.to(constants.DEVICE)
-            dm = get_data_model(config)
-            if os.path.isfile(f"{dm.model_name}_model.pt"):
-                os.remove(f"{dm.model_name}_model.pt")
-            run.file(f"{dm.model_name}_model.pt").download()
-            dm.load_data_model("")
+            dm, model_type = get_data_model(config)
+            if model_type == data_model.ModelType.Linear:
+                if os.path.isfile(f"{dm.model_name}_model.pt"):
+                    os.remove(f"{dm.model_name}_model.pt")
+                run.file(f"{dm.model_name}_model.pt").download()
+                dm.load_data_model("")
             return model_flow, dm, config
